Fix flaky test: testOnNetworkConnectivityReportedFalse
The test flake is likely caused by a carrier configuration
update. After the update is complete, the shell permission is
dropped, which causes the test network setup to fail due to a
lack of permission. The test network setup should also be
protected by a synchronized lock to avoid permission lost.

Bug: 296980394
Test: atest android.net.cts.ConnectivityDiagnosticsManagerTest \
      --iteration 50
Change-Id: I3c7a0a92cddeb7c0f41a11b929f72714f8b22c05