Cleanup existing firewall rule on myUid before the test
testFirewallCloseSocketAllowlistChainOtherUid sets the firewall rule on
myUid + 1 and confirms this does not affect the network access from
myUid.
This test failed when there is an existing firewall rule on myUid.
This CL cleans up the existing rule on myUid before running the test and
restores the existing state after the test.

Bug: 299898772
Test: atest CtsNetTestCases
Change-Id: I9083ac41a64a8b81a8b4616c0d57a25355f5d53c