Fix exception when unparcelling NetworkRequest with reservationId
This change adds a new private constructor that is called directly from
createFromParcel instead of using the public constructor used by CS.

The public constructor throws an exception when the passed
NetworkCapabilities specify a reservationId (to prevent apps from
passing in reservationIds). This means that it cannot be called from
createFromParcel.

Test: atest NetworkReservationTest
Change-Id: I1fe2411401623b0b01cbdbe040049b810fcbb48a