Add NetworkSecurityUnitTests to tethering-mainline-presubmit
Test with connectivity/tethering module only, to catch integration issues with older versions of other modules.

Only the unit test are added to the test_mapping for now. The CTS tests are not compatible with older dessert releases.

The min_sdk_version in the test build target should help preventing the
regressions from b/398885474.

Bug: 397130276
Test: on go/abtd
Change-Id: I215758a34901ff048a8279b4a417ef9f4797db0f