Hardcode the setting for multipath preference
Remove the dependency on hidden API by setting the constant in tests.

While there is a ConnectivitySettingsManager for this setting, it is
mostly broken (does not allow combining different bitmask values for the
multipath preference), and OEMs are much more likely to be setting the
setting directly. In practice the constant cannot be changed, otherwise
some of these callers may break. Therefore it is more valuable/realistic
to test behavior when the setting is modified directly.

Test: atest
Flag: TEST_ONLY
Bug: 410435307
Bug: 411565010
Change-Id: I802541ae0a6a6bab4b98c1a64dacb04775c349df