Test EthernetTracker#trackAvailableInterfaces
Coverage for this code path was removed a while ago, because the old
test relied on the old implementation of setEthernetEnabled(). This test
adds a similar test based on setIncludeTestInterfaces() instead, which
calls the trackAvailableInterfaces() code path.
trackAvailableInterfaces() is used when the ethernet service first
starts up (either at boot, or after a system server crash).

Test: TH
Flag: EXEMPT mainline
Bug: 1000001
Change-Id: I635031bb670ef62d4ccd0cb67d282ee907219089